removed paylaod data methods from individual tests and inserted in test helper
