+++ test/controllers/processing_request_test.rb
@@ -148,34 +148,6 @@ class ProcessingRequestTest < ControllerTest
                        rootUrl:     "http://id2.com" }
   end
 
-  def payload_data
-     {"payload" => {"url":"http://jumpstartlab.com/blog",
-      "requestedAt":"2013-02-16 21:38:28 -0700",
-      "respondedIn":37,
-      "referredBy":"http://jumpstartlab.com",
-      "requestType":"GET",
-      "parameters":[],
-      "eventName":"socialLogin",
-      "userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
-      "resolutionWidth":"1920",
-      "resolutionHeight":"1280",
-      "ip":"63.29.38.211" }.to_json }
-  end
-
-  def payload_data_two
-     {"payload" => {"url":"http://jumpstartlab.com/blog",
-      "requestedAt":"2013-02-16 21:38:28 -0700",
-      "respondedIn":666,
-      "referredBy":"http://jumpstartlab.com",
-      "requestType":"GET",
-      "parameters":[],
-      "eventName":"die with your boots on",
-      "userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
-      "resolutionWidth":"1920",
-      "resolutionHeight":"1280",
-      "ip":"63.29.38.211" }.to_json }
-  end
-
   def payload_data_id_2_com
     {"payload" => {"url":"http://something.com",
      "requestedAt":"2013-02-16 21:38:28 -0700",
