+++ test/controllers/viewing_data_and_statistics_test.rb
@@ -22,64 +22,8 @@ class ViewStatTest < ControllerTest
     post '/sources', { identifier:  "jumpstartlab",
                        rootUrl:     "http://jumpstartlab.com" }
     post '/sources/jumpstartlab/data', payload_data
-
     post '/sources/jumpstartlab/data', payload_data_four
     get "/sources/jumpstartlab/events/socialLogin"
   end
 
-  def payload_data
-     {"payload" => {"url":"http://jumpstartlab.com/blog",
-      "requestedAt":"2013-02-16 21:38:28 -0700",
-      "respondedIn":37,
-      "referredBy":"http://jumpstartlab.com",
-      "requestType":"GET",
-      "parameters":[],
-      "eventName":"socialLogin",
-      "userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
-      "resolutionWidth":"1920",
-      "resolutionHeight":"1280",
-      "ip":"63.29.38.211" }.to_json }
-  end
-
-  def payload_data_two
-     {"payload" => {"url":"http://jumpstartlab.com/blog",
-      "requestedAt":"2013-02-16 21:38:28 -0700",
-      "respondedIn":666,
-      "referredBy":"http://jumpstartlab.com",
-      "requestType":"GET",
-      "parameters":[],
-      "eventName":"die with your boots on",
-      "userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Opera/24.0.1309.0 Safari/537.17",
-      "resolutionWidth":"1920",
-      "resolutionHeight":"1280",
-      "ip":"63.29.38.211" }.to_json }
-  end
-
-  def payload_data_three
-     {"payload" => {"url":"http://jumpstartlab.com/test",
-      "requestedAt":"2013-02-16 21:38:28 -0700",
-      "respondedIn":666,
-      "referredBy":"http://fancyboots.com",
-      "requestType":"GET",
-      "parameters":[],
-      "eventName":"die with your boots on",
-      "userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
-      "resolutionWidth":"1920",
-      "resolutionHeight":"1280",
-      "ip":"63.29.38.211" }.to_json }
-  end
-
-  def payload_data_four
-     {"payload" => {"url":"http://jumpstartlab.com/blog",
-      "requestedAt":"2013-02-16 19:38:28 -0700",
-      "respondedIn":37,
-      "referredBy":"http://jumpstartlab.com",
-      "requestType":"GET",
-      "parameters":[],
-      "eventName":"socialLogin",
-      "userAgent":"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
-      "resolutionWidth":"1920",
-      "resolutionHeight":"1280",
-      "ip":"63.29.38.211" }.to_json }
-  end
 end
